requirements.txt
inference-sdk==0.9.17
opencv-python==4.10.0.84
google-generativeai==0.7.1
numpy
requests
fastapi
uvicorn